# api/openapi.yaml
openapi: 3.0.3
info:
  title: Office Tracker API
  description: |
    REST API for tracking office attendance using RFID cards.

    ## Authentication
    If API key authentication is enabled (via `API_KEY` environment variable),
    all endpoints except `/health` require authentication using one of these methods:
    - Header: `X-API-Key: your-api-key`
    - Header: `Authorization: Bearer your-api-key`

    ## CORS
    CORS can be enabled by setting the `CORS_ORIGINS` environment variable.
    - Set to `*` to allow all origins
    - Set to comma-separated list for specific origins (e.g., `http://localhost:3000,https://example.com`)
  version: 1.0.0
  contact:
    name: Office Tracker Team
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/users:
    get:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: List and search users
      description: Retrieve a list of users with optional search, sorting, and pagination
      tags:
        - Users
      parameters:
        - name: search
          in: query
          required: false
          description: Search users by name (partial match)
          schema:
            type: string
            example: John
        - name: limit
          in: query
          required: false
          description: Maximum number of results (default 50)
          schema:
            type: integer
            minimum: 1
            example: 50
        - name: offset
          in: query
          required: false
          description: Number of results to skip
          schema:
            type: integer
            minimum: 0
            example: 0
        - name: order
          in: query
          required: false
          description: Sort order (default 'asc')
          schema:
            type: string
            enum: [asc, desc]
            example: asc
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (default 'name')
          schema:
            type: string
            enum: [name, created_at]
            example: name
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Create a new user
      description: Register a new user in the system
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - {}
      summary: Delete users (bulk)
      description: |
        Delete users matching filters.
        Requires at least one filter (currently `search`) to prevent deleting all users.
      tags:
        - Users
      parameters:
        - name: search
          in: query
          required: false
          description: Delete users with matching names (partial match)
          schema:
            type: string
            example: John
        - name: order
          in: query
          required: false
          description: Sort order for filtered delete candidates
          schema:
            type: string
            enum: [asc, desc]
            example: asc
      responses:
        "200":
          description: Users deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                    format: int64
                    example: 3
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (optional, configured via API_KEY env var)
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication (alternative to X-API-Key)
  schemas:
...
    UserResponse:
      type: object
      description: User response DTO (excludes sensitive fields)
      properties:
        id:
          type: integer
          format: int64
          description: Unique user identifier
          example: 1
        name:
          type: string
          description: User's full name
          example: John Doe
        discord_id:
          type: string
          description: Discord user ID
          example: "123456789012345678"
      required:
        - id
        - name
        - discord_id
...
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
          description: User's full name
          example: John Doe
        rfid_uid:
          type: string
          description: RFID card unique identifier
          example: ABC123456
        discord_id:
          type: string
          description: Discord user ID (optional)
          example: "123456789012345678"
      required:
        - name
        - rfid_uid
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        text/plain:
          schema:
            type: string
            example: invalid request
...
    InternalServerError:
      description: Internal server error
      content:
        text/plain:
          schema:
            type: string
            example: database error